requirements.txt
# Core dependencies
python-dotenv==1.0.0
loguru==0.7.2
click==8.1.7

# Linear API
httpx==0.27.0
tenacity==8.2.3

# GitHub API  
PyGithub==2.2.0

# Development dependencies
pytest==8.0.0
pytest-asyncio==0.23.3
mypy==1.8.0 